hideAneLibSymbols no longer needs to be set
we released a major fix on all our iOS ANEs so you don't need to set
hideAneLibSymbols to yes in your project. we applied this fix because
xcode 7 had a bug which was causing AirSDK 20 or greater to act buggy
too! we couldn't wait for apple or Adobe to fix the problem so, we
applied a patch to all our ANEs to bypass the problem.